requirements.txt
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Menu==0.3.0
Flask-OAuthlib==0.9.1
Flask-SQLAlchemy==1.0
Flask-Uploads==0.1.3
Flask-WTF==0.9.5
Flask-WhooshAlchemy==0.55
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.8.1
SQLAlchemy==0.9.7
SQLAlchemy-Utils==0.30.1
Tempita==0.5.2
WTForms==1.0.5
WTForms-Alchemy==0.13.1
WTForms-Components==0.9.7
Werkzeug==0.10.4
Whoosh==2.6.0
alembic==0.7.5.post2
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
blinker==1.3
certifi==14.05.14
decorator==3.4.2
infinity==1.3
intervals==0.3.1
itsdangerous==0.24
oauthlib==0.7.2
pbr==0.10.8
redis==2.10.3
requests==2.6.1
requests-oauthlib==0.4.2
six==1.9.0
sqlalchemy-migrate==0.9.1
tornado==4.1
validators==0.7
wsgiref==0.1.2
